refactor(ai-employee): delete TS capability interfaces + portal customer.yaml editor
Removes ~3,800 lines of code with no in-product consumers:

src/lib/ai-employee/capabilities/ (entire dir, 14 files, ~2,300 lines)
  - 13 capability-interface files (Email, Calendar, InternalComms,
    DocumentStorage, PracticeManagement, ESign, CourtAccess, Payments,
    Accounting, IntakeCRM, CallTracking, conformance) defined typed
    contracts (EmailMessage, CalendarEvent, Attendee, ThreadQuery, …)
    that nothing imports except the test file that tested them.
  - Only CapabilityName (an 11-string union) was actually used by the
    customer.yaml validator. Inlined into customer-yaml/types.ts.
  - When a skill or adapter needs typed event/payload shapes for a
    real workflow, add the shape then.

src/lib/portal/ai-employee/customer-yaml-editor.ts (~680 lines)
src/pages/portal/products/ai-employee/settings/advanced/index.astro (~290 lines)
src/pages/api/portal/ai-employee/settings/customer-yaml-update.ts (~375 lines)
src/components/portal/ai-employee/customer-yaml-editor/*.astro (~500 lines)
tests/portal-customer-yaml-editor.test.ts (~400 lines)
  - Customer-side editor for customer.yaml — locked-field policy,
    projection, merger, diff, audit-metadata builder, principal-only
    page, HTML form POST handler.
  - Captain edits customer.yaml in git. No customer has portal access
    yet. The editor was built ahead of the audience that would use it.

tests/capability-conformance.test.ts
  - Tested the capability interfaces we just deleted.

Plus minor cleanup:
  - Removed the "Advanced: customer.yaml editor" link from the AI
    Employee settings landing (it pointed at the deleted page).
  - Removed the editor's exclusion in tests/forbidden-strings.test.ts
    (the page it documented as a non-list surface no longer exists).

Per the connectors audit (Captain decision 2026-05-26): kill code we
don't need now. Adapter and editor work returns when a paying customer
demands the workflow.
