--- src/components/portal/ai-employee/customer-yaml-editor/EscalationFields.astro
@@ -1,120 +1,0 @@
----
-/**
- * EscalationFields — form section for the escalation block.
- *
- * Escalation rosters control who receives red-flag alerts and failure
- * notifications from the agent. Recipients are email addresses; the
- * schema requires at least one of each. Acknowledgement window is
- * optional and defaults to 60 minutes upstream.
- *
- * Lists are entered as comma-separated emails — the server parses,
- * trims, deduplicates, and validates each entry. No client JS.
- */
-
-import type { EditableEscalation } from '../../../../lib/portal/ai-employee/customer-yaml-editor'
-import type { ValidationError } from '../../../../lib/ai-employee/customer-yaml'
-
-interface Props {
-  escalation: EditableEscalation
-  errors: ValidationError[]
-}
-
-const { escalation, errors } = Astro.props
-
-function errorsFor(path: string): ValidationError[] {
-  return errors.filter((e) => e.path === path || e.path.startsWith(`${path}.`))
-}
-
-const redFlagValue = escalation.red_flag_recipients.join(', ')
-const failureValue = escalation.failure_recipients.join(', ')
-const ackValue =
-  escalation.acknowledgement_window_minutes === null
-    ? ''
-    : String(escalation.acknowledgement_window_minutes)
----
-
-<section class="mt-12">
-  <header class="border-b-[2px] border-[color:var(--ss-color-text-primary)] pb-3 mb-6">
-    <p
-      class="font-['Archivo_Narrow'] text-label uppercase tracking-[0.18em] font-bold text-[color:var(--ss-color-primary)]"
-    >
-      § 02
-    </p>
-    <h2
-      class="mt-1 font-['Archivo'] text-xl md:text-2xl font-black uppercase tracking-[-0.02em] text-[color:var(--ss-color-text-primary)]"
-    >
-      Escalation
-    </h2>
-    <p class="mt-2 text-sm text-[color:var(--ss-color-text-secondary)]">
-      Who gets notified when the agent flags a red-flag pattern or hits a failure mode.
-    </p>
-  </header>
-
-  <div class="space-y-6">
-    <label class="block">
-      <span
-        class="font-['Archivo_Narrow'] text-label uppercase tracking-[0.14em] font-bold text-[color:var(--ss-color-text-secondary)]"
-      >
-        Red-flag recipients
-      </span>
-      <input
-        type="text"
-        name="escalation.red_flag_recipients"
-        value={redFlagValue}
-        required
-        placeholder="alerts@firm.com, owner@firm.com"
-        class="mt-2 w-full px-4 py-2 border-[3px] border-[color:var(--ss-color-text-primary)] bg-[color:var(--ss-color-background)] text-[color:var(--ss-color-text-primary)] focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[color:var(--ss-color-action)] focus-visible:ring-offset-2"
-      />
-      <span class="mt-1 inline-block text-xs text-[color:var(--ss-color-text-muted)]">
-        Comma-separated emails. At least one required.
-      </span>
-    </label>
-
-    <label class="block">
-      <span
-        class="font-['Archivo_Narrow'] text-label uppercase tracking-[0.14em] font-bold text-[color:var(--ss-color-text-secondary)]"
-      >
-        Failure recipients
-      </span>
-      <input
-        type="text"
-        name="escalation.failure_recipients"
-        value={failureValue}
-        required
-        placeholder="alerts@firm.com, owner@firm.com"
-        class="mt-2 w-full px-4 py-2 border-[3px] border-[color:var(--ss-color-text-primary)] bg-[color:var(--ss-color-background)] text-[color:var(--ss-color-text-primary)] focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[color:var(--ss-color-action)] focus-visible:ring-offset-2"
-      />
-      <span class="mt-1 inline-block text-xs text-[color:var(--ss-color-text-muted)]">
-        Comma-separated emails. At least one required.
-      </span>
-    </label>
-
-    <label class="block max-w-xs">
-      <span
-        class="font-['Archivo_Narrow'] text-label uppercase tracking-[0.14em] font-bold text-[color:var(--ss-color-text-secondary)]"
-      >
-        Acknowledgement window (minutes)
-      </span>
-      <input
-        type="number"
-        name="escalation.acknowledgement_window_minutes"
-        value={ackValue}
-        min="1"
-        max="1440"
-        placeholder="60"
-        class="mt-2 w-full px-4 py-2 border-[3px] border-[color:var(--ss-color-text-primary)] bg-[color:var(--ss-color-background)] text-[color:var(--ss-color-text-primary)] focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[color:var(--ss-color-action)] focus-visible:ring-offset-2"
-      />
-      <span class="mt-1 inline-block text-xs text-[color:var(--ss-color-text-muted)]">
-        Leave blank for the default of 60 minutes.
-      </span>
-    </label>
-
-    {
-      errorsFor('escalation').map((err) => (
-        <p class="text-sm text-[color:var(--ss-color-error)]">
-          <span class="font-['JetBrains_Mono']">{err.path}</span>: {err.message}
-        </p>
-      ))
-    }
-  </div>
-</section>
