--- src/components/portal/ai-employee/customer-yaml-editor/ScopeFields.astro
@@ -1,106 +1,0 @@
----
-/**
- * ScopeFields — form section for the scope envelope.
- *
- * Scope controls email-folder visibility, keyword blocks, domain blocks,
- * and matter blocks. The agent reads only inside the visibility set
- * minus the blind set; messages matching blocked keywords or originating
- * from blocked domains never reach the agent's view. Matter blocks are
- * an optional list of external PM matter refs the agent must not touch.
- *
- * All lists are entered as comma-separated values. The server parses,
- * trims, and deduplicates each entry. No client JS.
- */
-
-import type { EditableScope } from '../../../../lib/portal/ai-employee/customer-yaml-editor'
-import type { ValidationError } from '../../../../lib/ai-employee/customer-yaml'
-
-interface Props {
-  scope: EditableScope
-  errors: ValidationError[]
-}
-
-const { scope, errors } = Astro.props
-
-function errorsFor(path: string): ValidationError[] {
-  return errors.filter((e) => e.path === path || e.path.startsWith(`${path}.`))
-}
-
-const FIELDS = [
-  {
-    name: 'email_folders_visible',
-    label: 'Visible email folders',
-    hint: 'Comma-separated folder names the agent may read.',
-  },
-  {
-    name: 'email_folders_blind',
-    label: 'Blind email folders',
-    hint: 'Comma-separated folder names the agent must NOT read.',
-  },
-  {
-    name: 'email_keyword_blocks',
-    label: 'Email keyword blocks',
-    hint: 'Comma-separated keywords; matching messages are hidden from the agent.',
-  },
-  {
-    name: 'domain_blocks',
-    label: 'Domain blocks',
-    hint: 'Comma-separated domains; matching senders are hidden from the agent.',
-  },
-  {
-    name: 'matter_blocks',
-    label: 'Matter blocks',
-    hint: 'Comma-separated external PM matter refs the agent must not touch.',
-  },
-] as const
----
-
-<section class="mt-12">
-  <header class="border-b-[2px] border-[color:var(--ss-color-text-primary)] pb-3 mb-6">
-    <p
-      class="font-['Archivo_Narrow'] text-label uppercase tracking-[0.18em] font-bold text-[color:var(--ss-color-primary)]"
-    >
-      § 05
-    </p>
-    <h2
-      class="mt-1 font-['Archivo'] text-xl md:text-2xl font-black uppercase tracking-[-0.02em] text-[color:var(--ss-color-text-primary)]"
-    >
-      Scope envelope
-    </h2>
-    <p class="mt-2 text-sm text-[color:var(--ss-color-text-secondary)]">
-      Visibility and block lists that bound what the agent may see and act on.
-    </p>
-  </header>
-
-  <div class="space-y-6">
-    {
-      FIELDS.map((field) => {
-        const value = scope[field.name as keyof EditableScope].join(', ')
-        return (
-          <label class="block">
-            <span class="font-['Archivo_Narrow'] text-label uppercase tracking-[0.14em] font-bold text-[color:var(--ss-color-text-secondary)]">
-              {field.label}
-            </span>
-            <input
-              type="text"
-              name={`scope.${field.name}`}
-              value={value}
-              class="mt-2 w-full px-4 py-2 border-[3px] border-[color:var(--ss-color-text-primary)] bg-[color:var(--ss-color-background)] text-[color:var(--ss-color-text-primary)] focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[color:var(--ss-color-action)] focus-visible:ring-offset-2"
-            />
-            <span class="mt-1 inline-block text-xs text-[color:var(--ss-color-text-muted)]">
-              {field.hint}
-            </span>
-          </label>
-        )
-      })
-    }
-
-    {
-      errorsFor('scope').map((err) => (
-        <p class="text-sm text-[color:var(--ss-color-error)]">
-          <span class="font-['JetBrains_Mono']">{err.path}</span>: {err.message}
-        </p>
-      ))
-    }
-  </div>
-</section>
