+++ src/pages/portal/products/ai-employee/settings/index.astro
@@ -161,16 +161,6 @@ const banner: { kind: 'ok' | 'error'; text: string } | null = (() => {
       <SkillTogglesSection rows={view.skillToggleRows} personaSlug={view.personaSlug} />
 
       <ConnectorStatusSection rows={view.connectorRows} />
-
-      <section class="mt-section">
-        <a
-          href="/portal/products/ai-employee/settings/customer-yaml"
-          class="inline-flex items-center gap-2 min-h-11 font-['Archivo_Narrow'] text-xs md:text-sm font-bold uppercase tracking-[0.14em] text-[color:var(--ss-color-primary)] hover:text-[color:var(--ss-color-text-primary)] focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[color:var(--ss-color-action)] focus-visible:ring-offset-2"
-        >
-          Advanced: customer.yaml editor
-          <span aria-hidden="true" class="font-['Archivo'] font-black">→</span>
-        </a>
-      </section>
     </main>
   </body>
 </html>
